@@ -1,6 +1,6 @@
 {
   "name": "fball_ui",
-  "version": "0.1.0",
+  "version": "1.0.2",
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
@@ -10,17 +10,15 @@
   "dependencies": {
     "@babel/polyfill": "^7.11.5",
     "axios": "^1.7.4",
-    "bootstrap": "^4.6.0",
-    "bootstrap-vue": "^2.23.1",
+    "bootstrap": "^5.3.2",
+    "bootstrap-vue-next": "^0.30.3",
     "core-js": "^3.6.5",
     "luxon": "^2.1.1",
     "popper.js": "^1.16.1",
-    "portal-vue": "^2.1.7",
     "underscore": "^1.13.1",
-    "vue": "^2.6.11",
-    "vue-luxon": "^0.10.0",
-    "vue-router": "^3.6.5",
-    "vuex": "^3.6.2"
+    "vue": "^3.5.0",
+    "vue-router": "^4.2.5",
+    "vuex": "^4.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.16",
@@ -31,16 +29,15 @@
     "@vue/cli-plugin-router": "~5.0.0",
     "@vue/cli-plugin-vuex": "~5.0.0",
     "@vue/cli-service": "~5.0.8",
+    "@vue/compiler-sfc": "^3.4.15",
     "@vue/runtime-dom": "^3.2.47",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-vue": "^8.0.3",
     "mutationobserver-shim": "^0.3.7",
     "prettier": "^3.1.1",
-    "sass": "^1.80.3",
-    "sass-loader": "^10.5.2",
-    "vue-cli-plugin-bootstrap-vue": "~0.8.2",
-    "vue-loader": "^17.4.2",
-    "vue-template-compiler": "^2.6.14"
+    "sass": "^1.89.2",
+    "sass-loader": "^16.0.5",
+    "vue-loader": "^17.4.2"
   },
   "eslintConfig": {
     "root": true,
@@ -48,7 +45,7 @@
       "node": true
     },
     "extends": [
-      "plugin:vue/essential",
+      "plugin:vue/vue3-essential",
       "eslint:recommended"
     ],
     "parserOptions": {
