@@ -1,17 +1,17 @@
 import "@babel/polyfill";
 import "mutationobserver-shim";
-import Vue from "vue";
-import "./plugins/bootstrap-vue";
+import { createApp } from "vue";
+import bootstrapVue from "./plugins/bootstrap-vue";
+import "bootstrap/dist/css/bootstrap.css";
+import "bootstrap-vue-next/dist/bootstrap-vue-next.css";
 import App from "./App.vue";
-Vue.config.productionTip = false;
-import VueLuxon from "vue-luxon";
 import router from "./router";
 import store from "./store";
 
-Vue.use(VueLuxon);
+const app = createApp(App);
 
-new Vue({
-  router,
-  store,
-  render: (h) => h(App),
-}).$mount("#app");
+app.use(bootstrapVue);
+app.use(router);
+app.use(store);
+
+app.mount("#app");
