@@ -1,12 +1,11 @@
-const bootstrapSassAbstractsImports = require("vue-cli-plugin-bootstrap-vue/sassAbstractsImports.js");
 module.exports = {
   css: {
     loaderOptions: {
       sass: {
-        additionalData: bootstrapSassAbstractsImports.join("\n"),
+        additionalData: "",
       },
       scss: {
-        additionalData: [...bootstrapSassAbstractsImports, ""].join(";\n"),
+        additionalData: "",
       },
     },
   },
